Refactor utp into multiple files
