+++ llarp/CMakeLists.txt
@@ -170,6 +170,9 @@ set(LIB_SRC
   link/iwp.cpp
   link/server.cpp
   link/session.cpp
+  link/utp_inbound_message.cpp
+  link/utp_linklayer.cpp
+  link/utp_session.cpp
   link/utp.cpp
   messages/dht_immediate.cpp
   messages/dht.cpp
